--- validations/email.go
@@ -1,8 +1,0 @@
-package validations
-
-import "regexp"
-
-func IsValidEmail(email string) bool {
-	regex := regexp.MustCompile(`^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$`)
-	return regex.MatchString(email)
-}
